Added second fragment, and shared viewmodel. Created functions to get BookPrices and BookOrders from repo.
