--- app/src/main/java/dev/ricsarabia/cryptochallenge/ui/detail/DetailViewModel.kt
@@ -1,7 +1,0 @@
-package dev.ricsarabia.cryptochallenge.ui.detail
-
-import androidx.lifecycle.ViewModel
-
-class DetailViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
