+++ app/src/main/res/layout/detail_fragment.xml
@@ -8,6 +8,6 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="Hello" />
+        android:text="Hello from detail" />
 
 </FrameLayout>
